Fixed things following review (part 2)
