+++ tests/types.ts
@@ -1,0 +1,14 @@
+export interface Character {
+  name: string;
+  role: string;
+  birth: number;
+}
+
+export interface SimpleCharacterKey {
+  name: string;
+}
+
+export interface ComplexCharacterKey {
+  name: string;
+  birth: number;
+}
